requirements.txt
dnspython==2.7.0
requests==2.32.3
pytest==8.3.4
pytest-timeout==2.3.1
